+++ packages/orq-rc/src/orq_ai_sdk/models/createchunkop.py
@@ -257,7 +257,7 @@ class CreateChunkResponseBody(BaseModel):
     updated_by_id: str
     r"""The unique identifier of the user who updated the chunk"""
 
-    id: Optional[str] = "chunk_01JQTR12W8AX8P0X835MHV8JH9"
+    id: Optional[str] = "chunk_01JQTTXB133A23MQBW96KPPXMW"
     r"""Unique identifier for the element"""
 
     enabled: Optional[bool] = True