+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQTR12W6ZPXDKAK5YXD7KW8K"
+        "01JQTTXB11QF8M06BKM4E49WJ0"
     )
     r"""The id of the resource"""