+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -60,5 +60,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-04-02T07:56:32.815Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-04-02T08:46:55.819Z")
     r"""The date and time the resource was created"""