+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQTR12W54M2AA20KVEBYYFDB"
+        "01JQTTXB10M2VYPJZFQMEE4D25"
     )
     r"""The id of the resource"""