+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -79,5 +79,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-02T07:56:31.295Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-02T08:46:54.224Z")
     r"""The date and time the resource was last updated"""