+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQTR12W89P8KYC5W37XYYMPB"
+        "01JQTTXB13GNAX2WJ2CQZA8SNC"
     )
     r"""The id of the resource"""