+++ packages/orq-rc/RELEASES.md
@@ -1069,3 +1069,13 @@ Based on:
 - [python v3.3.0-rc.25] packages/orq-rc
 ### Releases
 - [PyPI v3.3.0-rc.25] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.25 - packages/orq-rc
+
+## 2025-04-02 08:48:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.527.1 (2.564.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0-rc.26] packages/orq-rc
+### Releases
+- [PyPI v3.3.0-rc.26] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.26 - packages/orq-rc